requirements.txt
asgiref==3.7.2
attrs==23.1.0
azure-core==1.27.1
azure-storage-blob==12.16.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
culqi==1.0.1
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
jsonschema==4.17.3
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
